Add event listener to submit button
Submit button reads input values and creates a new book object with
those values. Then empties the modal input elements, toggles the modal,
adds the new book to the library array, and finally posts the book
to the DOM. Added forEach method to post books function to clear the
DOM to allow for later implementation of index iteration, which will
allow delete button to read the index of the book and remove it from
the array.